Graph vis package refactor v3
