--- src/framework/graph-visualiser/data/index.ts
@@ -1,3 +1,0 @@
-export * from "./types";
-export { constructGraphFromRowsResult } from "./builder";
-export * from "./back-compat";
