+++ src/framework/graph-visualiser/data/types.ts
@@ -23,6 +23,12 @@ export type DataVertexSpecial = VertexUnavailable | VertexFunction | VertexExpre
 export type DataVertexKind = ThingKind | TypeKind | ValueKind | SpecialVertexKind;
 export type DataVertex = Concept | DataVertexSpecial;
 
+export function getTypeLabel(vertex: DataVertex): string | undefined {
+    if ("type" in vertex && vertex.type && "label" in vertex.type) return vertex.type.label;
+    if ("label" in vertex && vertex.kind !== "unavailable") return vertex.label;
+    return undefined;
+}
+
 export type QueryCoordinates = { branch: number, constraint: number };
 
 export type DataGraph = {
