+++ src/framework/graph-visualiser/node-programs/diamond/index.ts
