+++ src/framework/graph-visualiser/node-programs/diamond/shader-frag.ts
