+++ src/framework/graph-visualiser/node-programs/diamond/shader-vert.ts
