+++ src/framework/graph-visualiser/node-programs/diamond/utils.ts
@@ -1,9 +1,9 @@
 import { Attributes } from "graphology-types";
 import { Settings } from "sigma/settings";
 import { NodeDisplayData, PartialButFor } from "sigma/types";
-import { drawCenteredNodeLabel } from "../label-utils";
+import { drawCenteredNodeLabel } from "../../sigma-label-utils";
 
-export function drawRoundedRectNodeLabel<
+export function drawDiamondNodeLabel<
     N extends Attributes = Attributes,
     E extends Attributes = Attributes,
     G extends Attributes = Attributes,
@@ -15,7 +15,7 @@ export function drawRoundedRectNodeLabel<
     return drawCenteredNodeLabel<N, E, G>(context, data, settings);
 }
 
-export function drawRoundedRectNodeHover<
+export function drawDiamondNodeHover<
     N extends Attributes = Attributes,
     E extends Attributes = Attributes,
     G extends Attributes = Attributes,
@@ -35,19 +35,14 @@ export function drawRoundedRectNodeHover<
     const rawW = (data as any).width ?? data.size;
     const rawH = (data as any).height ?? data.size;
     const scale = data.size / Math.max(rawW, rawH);
-    const radiusX = rawW * scale + PADDING;
-    const radiusY = rawH * scale + PADDING;
-    const cornerR = Math.min(radiusX, radiusY) * 0.25;
+    const rx = rawW * scale + PADDING;
+    const ry = rawH * scale + PADDING;
 
     context.beginPath();
-    context.moveTo(data.x + radiusX - cornerR, data.y - radiusY);
-    context.arcTo(data.x + radiusX, data.y - radiusY, data.x + radiusX, data.y - radiusY + cornerR, cornerR);
-    context.lineTo(data.x + radiusX, data.y + radiusY - cornerR);
-    context.arcTo(data.x + radiusX, data.y + radiusY, data.x + radiusX - cornerR, data.y + radiusY, cornerR);
-    context.lineTo(data.x - radiusX + cornerR, data.y + radiusY);
-    context.arcTo(data.x - radiusX, data.y + radiusY, data.x - radiusX, data.y + radiusY - cornerR, cornerR);
-    context.lineTo(data.x - radiusX, data.y - radiusY + cornerR);
-    context.arcTo(data.x - radiusX, data.y - radiusY, data.x - radiusX + cornerR, data.y - radiusY, cornerR);
+    context.moveTo(data.x, data.y - ry);
+    context.lineTo(data.x + rx, data.y);
+    context.lineTo(data.x, data.y + ry);
+    context.lineTo(data.x - rx, data.y);
     context.closePath();
     context.fill();
 
@@ -55,5 +50,5 @@ export function drawRoundedRectNodeHover<
     context.shadowOffsetY = 0;
     context.shadowBlur = 0;
 
-    drawRoundedRectNodeLabel(context, data, settings);
+    drawDiamondNodeLabel(context, data, settings);
 }
