+++ src/framework/graph-visualiser/node-programs/ellipse/index.ts
