+++ src/framework/graph-visualiser/node-programs/ellipse/shader-frag.ts
