+++ src/framework/graph-visualiser/node-programs/ellipse/shader-vert.ts
