+++ src/framework/graph-visualiser/node-programs/ellipse/utils.ts
@@ -1,9 +1,9 @@
 import { Attributes } from "graphology-types";
 import { Settings } from "sigma/settings";
 import { NodeDisplayData, PartialButFor } from "sigma/types";
-import { drawCenteredNodeLabel } from "../label-utils";
+import { drawCenteredNodeLabel } from "../../sigma-label-utils";
 
-export function drawDiamondNodeLabel<
+export function drawEllipseNodeLabel<
     N extends Attributes = Attributes,
     E extends Attributes = Attributes,
     G extends Attributes = Attributes,
@@ -15,7 +15,7 @@ export function drawDiamondNodeLabel<
     return drawCenteredNodeLabel<N, E, G>(context, data, settings);
 }
 
-export function drawDiamondNodeHover<
+export function drawEllipseNodeHover<
     N extends Attributes = Attributes,
     E extends Attributes = Attributes,
     G extends Attributes = Attributes,
@@ -35,14 +35,11 @@ export function drawDiamondNodeHover<
     const rawW = (data as any).width ?? data.size;
     const rawH = (data as any).height ?? data.size;
     const scale = data.size / Math.max(rawW, rawH);
-    const rx = rawW * scale + PADDING;
-    const ry = rawH * scale + PADDING;
+    const radiusX = rawW * scale + PADDING;
+    const radiusY = rawH * scale + PADDING;
 
     context.beginPath();
-    context.moveTo(data.x, data.y - ry);
-    context.lineTo(data.x + rx, data.y);
-    context.lineTo(data.x, data.y + ry);
-    context.lineTo(data.x - rx, data.y);
+    context.ellipse(data.x, data.y, radiusX, radiusY, 0, 0, Math.PI * 2);
     context.closePath();
     context.fill();
 
@@ -50,5 +47,5 @@ export function drawDiamondNodeHover<
     context.shadowOffsetY = 0;
     context.shadowBlur = 0;
 
-    drawDiamondNodeLabel(context, data, settings);
+    drawEllipseNodeLabel(context, data, settings);
 }
