+++ src/framework/graph-visualiser/node-programs/rounded-rect/index.ts
