+++ src/framework/graph-visualiser/node-programs/rounded-rect/shader-frag.ts
