--- src/framework/graph-visualiser/rendering/index.ts
@@ -1,5 +1,0 @@
-export { drawCenteredNodeLabel, setUseBorderColorForLabels, zoomScaledFontSize } from "./label-utils";
-export { defaultSigmaSettings, createSigmaRenderer } from "./sigma-settings";
-export { NodeDiamondProgram } from "./node-diamond";
-export { NodeEllipseProgram } from "./node-ellipse";
-export { NodeRoundedRectangleProgram } from "./node-rounded-rect";
