+++ src/framework/graph-visualiser/sigma-label-utils.ts
