--- src/framework/graph-visualiser/style/index.ts
@@ -1,12 +1,0 @@
-export {
-    StudioConverterStyleParameters,
-    darkPalette,
-    defaultEdgeLabelColors,
-    defaultQueryStyleParameters,
-    defaultExplorationQueryStyleParameters,
-} from "./parameters";
-export {
-    colorEdgesByConstraintIndex,
-    colorQuery,
-    getColorForConstraintIndex,
-} from "./constraint-colors";
