--- src/framework/graph-visualiser/visual/index.ts
@@ -1,9 +1,0 @@
-export * from "./types";
-export {
-    ILogicalGraphConverter,
-    convertLogicalGraphWith,
-    StudioConverter,
-    shouldCreateNode,
-    shouldCreateEdge,
-    vertexMapKey,
-} from "./converter";
