+++ src/service/schema-state.service.ts
@@ -13,8 +13,8 @@ import Graph from "graphology";
 import { BehaviorSubject, combineLatest, distinctUntilChanged, finalize, first, map } from "rxjs";
 import Sigma, { Camera } from "sigma";
 import { GraphVisualiser } from "../framework/graph-visualiser";
-import { createSigmaRenderer, defaultSigmaSettings } from "../framework/graph-visualiser/rendering/sigma-settings";
-import { newVisualGraph } from "../framework/graph-visualiser/visual/types";
+import { createSigmaRenderer, defaultSigmaSettings } from "../framework/graph-visualiser/sigma-settings";
+import { newVisualGraph } from "../framework/graph-visualiser/visual-graph";
 import { Layouts } from "../framework/graph-visualiser/layout";
 import { DriverState } from "./driver-state.service";
 import { GraphStyleService } from "./graph-style.service";
